requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
banal==1.0.6
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
bleach==6.2.0
Brotli==1.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
controlnet_aux==0.0.10
cycler==0.12.1
dataset==1.6.2
datasets==3.6.0
deepspeed==0.17.5
diffuser==0.0.1
# Editable Git install with no remote (diffusers==0.35.0.dev0)
dill==0.3.8
einops==0.8.1
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
hf-xet==1.1.4
hjson==3.1.0
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
inflate64==1.0.3
Jinja2==3.1.6
kaggle==1.7.4.5
kagglehub==0.3.12
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
multiprocess==0.70.16
multivolumefile==0.2.3
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optimum==1.26.1
packaging==25.0
pandas==2.3.0
peft==0.17.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
py7zr==1.0.0
pyarrow==20.0.0
pybcj==1.0.6
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyppmd==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzstd==0.17.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.30.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
soupsieve==2.7
SQLAlchemy==1.4.54
sympy==1.14.0
text-unidecode==1.3
texttable==1.7.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.22.0
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.56.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wandb==0.20.1
webencodings==0.5.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
